expose the account in signAccountOp.ts; move getTokenUsdAmount to helper.ts;
