+++ src/controllers/signAccountOp/helper.ts
@@ -1,6 +1,7 @@
-import { ZeroAddress } from 'ethers'
+import { formatUnits, ZeroAddress } from 'ethers'
 
 import { FeePaymentOption } from '../../libs/estimate/interfaces'
+import { Price, TokenResult } from '../../libs/portfolio'
 
 export function getFeeSpeedIdentifier(option: FeePaymentOption, accountAddr: string) {
   // if the token is native and we're paying with EOA, we do not need
@@ -12,4 +13,20 @@ export function getFeeSpeedIdentifier(option: FeePaymentOption, accountAddr: str
   return `${paidBy}:${option.token.address}:${
     option.token.flags.onGasTank ? 'gasTank' : 'feeToken'
   }`
+}
+
+export function getTokenUsdAmount(token: TokenResult, gasAmount: bigint): string {
+  const isUsd = (price: Price) => price.baseCurrency === 'usd'
+  const usdPrice = token.priceIn.find(isUsd)?.price
+
+  if (!usdPrice) return ''
+
+  const usdPriceFormatted = BigInt(usdPrice * 1e18)
+
+  console.log('THE GAS AMOUNT')
+  console.log(gasAmount)
+  console.log('THE GAS AMOUNT')
+
+  // 18 it's because we multiply usdPrice * 1e18 and here we need to deduct it
+  return formatUnits(gasAmount * usdPriceFormatted, 18 + token.decimals)
 }
